Support: Date, Boolean value & refactoring E2E
