+++ e2e/specs/content_managemant_api/entry.spec
@@ -27,11 +27,17 @@ tags: default
 * MongoDBの"ENTRY"に登録されている値のJsonPath"$.items[0].value"の数値が"100"である
 
 ## Entryのフィールドに日時を保存することができる
-tags: unimplemented
 * "/v1/spaces/spaceId/entries"にボディ"setup/request/entry_with_date.json"でPOSTリクエストを送る
 * "201"ステータスコードが返ってくる
 * MongoDBの"ENTRY"に登録されている値のJsonPath"$.content_model_id"の値が"5063114bd386d8fadbd6b008"である
 * MongoDBの"ENTRY"に登録されている値のJsonPath"$.items[0].value"の値が"2021-08-02T19:46:00Z"である
+
+## Entryのフィールドに真偽値を保存することができる
+* "/v1/spaces/spaceId/entries"にボディ"setup/request/entry_with_boolean.json"でPOSTリクエストを送る
+* "201"ステータスコードが返ってくる
+* MongoDBの"ENTRY"に登録されている値のJsonPath"$.content_model_id"の値が"5063114bd386d8fadbd6b009"である
+* MongoDBの"ENTRY"に登録されている値のJsonPath"$.items[0].value"の真偽値が"true"である
+* MongoDBの"ENTRY"に登録されている値のJsonPath"$.items[1].value"の真偽値が"false"である
 
 ## EntryがContentModelの形に沿っていない場合は400エラーを返す
 * "/v1/spaces/spaceId/entries"にボディ"setup/request/entry_unmatch_to_content_model.json"でPOSTリクエストを送る
