+++ e2e/src/test/resources/setup/request/entry_with_boolean.json
@@ -1,0 +1,11 @@
+{
+  "content-model-id": "5063114bd386d8fadbd6b009",
+  "items": [
+    {
+      "value": true
+    },
+    {
+      "value": false
+    }
+  ]
+}
